deploy_hf_model/requirements.txt
torch>=2.0.0
transformers>=4.30.0
huggingface_hub>=0.23.0
Pillow>=9.5.0
